Update User.php
Ensure standard name output with capitalization. I haven't looked at encoding to see how it handles UTF-8, but should work. May need to a check for the current language and ensure it should *actually* use `ucwords`, but I'm not entirely sure what languages are appropriate or not, but it's usually standard for most languages, including English.