Detect dataType from server response in AJAX requests
Previously: if the AJAX query doesn't contain a dataType, we
assume the dataType is "html".

Now: if the AJAX query doesn't contain a dataType, we determine the
dataType from the server's content-type (either "html" or "json").

This is is particularly useful in `modal.load(url)` scenarios and
other places where the response type can be both "html" and "json".