improve performance for search (avoid memory issues)
when limiting a search query to spaces, only the space IDs are necessary.
When a user is member in many spaces, loading all the spaces objects
consumes a lot of memory and in extreme cases can cause Humhub to crash
with out of memory error.

This problem is fixed by adding a new method
`Membership::GetUserSpaceIds()` a new private method is added to ensure
the query is the same for `GetUserSpaceIds()` and `GetUserSpaces()`.