Fix profile Friends panel id
The profile Friends panel was using the same div id as the Followers
panel, causing it to collapse at the same time as the Followers panel,
and preventing it from collapsing from its own widget panel menu.
Fix this by correcting the div id to match the panel widget
configuration.