Allow current space visibility to remain
Include the current space visibility among the available options, even
if the current user lacks permission to use it in a new space. This
could happen if another space admin with appropriate permission has set
the visibility.

The space model validation allows the visibility to remain unchanged
even if it would otherwise fail validation.