Create file_versioning.sql
Hello Lucas,

with this there could be implemented an File Versioning in Humhub, directly in the database.

The following things needs to be done:
A) Interface for Users                  I am not able to implement any interface or Javascript things :-( 
B) Integrate it in the files module itself

I hope my work will help your nice product. 

And please I need a reward. This is (I think) the wrong way to give code back. But I do not know how to make it right in the github world. 

With best regards

Rainer Garbotz