Update HH Branding (#7649)
* Update HH Branding

Idk if the HumHub logo in the installer is styled somewhere in CSS, but the old logo was 200x60px and the new one is 843x248px, so if it’s not limited via CSS, the logo might be gigantic. I changed the size to 248px so it’s more "HD" than the current one.

https://github.com/humhub/humhub-internal/issues/718

* Resize Logo + Honeycomb

* Image Size -> CSS

* 248 -> 296